// src/models/entities/receipt.ts
interface Receipt {
    receiptNumber: string
    date: string
    customerName: string
    phone: string
    address: string
    lightDescription: string
    lightPrice: string
    soundDescription: string
    soundPrice: string
    tvDescription: string
    tvPrice: string
    stageDescription: string
    stagePrice: string
    totalPrice: string
    discount: string
    finalPrice: string
    customerPayment: string
    customerDebt: string
}

export default Receipt